More work on deteciton plot.
Work on the detection plot for allowing the plotting of waveforms at
mulitple scales.

+++ AiPAM/src/main/java/com/jamdev/maven/aipam/layout/detectionDisplay/DataProvider.java
@@ -47,7 +47,7 @@ public interface DataProvider {
 	 * @param start - the start in millis date number.
 	 * @param end   - the end in millis date number.
 	 */
-	public void requestData(long start, long end);
+	public void requestData(double start, double end);
 
 	public DataLimits getDataLimits();
